requirements.txt
requests
selenium
ipapi
undetected-chromedriver
selenium-wire
pytest-playwright